bosch_database/requirements.txt
pymongo==3.11.4
paho-mqtt